requirements.txt
anyio==3.4.0
asgiref==3.4.1
click==8.0.3
fastapi==0.71.0
h11==0.12.0
idna==3.3
Pillow==9.0.0
pydantic==1.9.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.0.1
uvicorn==0.16.0
